+++ lib/components/AppLoading.js
@@ -1,7 +1,7 @@
 /* globals __BUILD__ */
 import React from 'react'
 import Base from './Base'
-import Spinner from './Spinner'
+import Spinner from './chrome/Spinner'
 import AppDrag from './AppDrag'
 
 import { defaultPalette } from '../propTypes/palette'
