+++ lib/components/Datasets.js
@@ -13,9 +13,9 @@ import SearchResultsContainer from '../containers/SearchResults'
 
 import List from './List'
 import DatasetItem from './item/DatasetItem'
-import Spinner from './Spinner'
+import Spinner from './chrome/Spinner'
 import Header from './Header'
-import Button from './Button'
+import Button from './chrome/Button'
 import { addActiveToLink } from '../utils/links.js'
 
 class Datasets extends Base {
