+++ lib/components/Header.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types'
 import Base from './Base'
 import NavLinks from './NavLinks'
 import { defaultPalette } from '../propTypes/palette'
-import Spinner from './Spinner'
+import Spinner from './chrome/Spinner'
 
 export default class Header extends Base {
   template (css) {
