+++ lib/components/List.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import LoadMore from './LoadMore'
+import LoadMore from './chrome/LoadMore'
 
 function selectFunc (fn, data, i) {
   return () => {
