+++ lib/components/Profiles.js
@@ -6,7 +6,7 @@ import Base from './Base'
 import ReadOnly from './ReadOnly'
 import List from './List'
 import ProfileItem from './item/ProfileItem'
-import Spinner from './Spinner'
+import Spinner from './chrome/Spinner'
 
 export default class Profiles extends Base {
   constructor (props) {
