+++ lib/components/TabPanel.js
@@ -13,8 +13,7 @@ export default class TabPanel extends Base {
     };
 
     [
-      'handleSetPanel',
-      'setLocationBarText'
+      'handleSetPanel'
     ].forEach((m) => { this[m] = this[m].bind(this) })
   }
