+++ lib/components/chrome/Button.js
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom'
 import PropTypes from 'prop-types'
 
 import Spinner from './Spinner'
-import Base from './Base'
+import Base from '../Base'
 
 export default class Button extends Base {
   template (css) {
