+++ lib/components/chrome/Dropdown.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import Base from './Base'
+import Base from '../Base'
 import Button from './Button'
 
 export default class Dropdown extends Base {
