+++ lib/components/chrome/LoadMore.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import Base from './Base'
+import Base from '../Base'
 import Spinner from './Spinner'
 import Button from './Button'
