+++ lib/components/chrome/Spinner.js
@@ -1,8 +1,8 @@
 import React from 'react'
-import Base from './Base'
+import Base from '../Base'
 import PropTypes from 'prop-types'
 
-import { defaultPalette } from '../propTypes/palette'
+import { defaultPalette } from '../../propTypes/palette'
 
 export default class Spinner extends Base {
   template (css) {
