+++ lib/components/dataset/Body.js
@@ -3,9 +3,9 @@ import PropTypes from 'prop-types'
 
 import Base from '../Base'
 import Json from '../Json'
-import Spinner from '../Spinner'
+import Spinner from '../chrome/Spinner'
 import DatasetRefProps from '../../propTypes/datasetRefProps'
-import Button from '../Button'
+import Button from '../chrome/Button'
 import HandsonTable from '../HandsonTable'
 import { extractSchema } from '../../selectors/dataset'
 import BodyReadOnly from './BodyReadOnly'
