+++ lib/components/dataset/Commits.js
@@ -5,7 +5,7 @@ import { debounce } from 'lodash'
 import Base from '../Base'
 import List from '../List'
 import CommitItemContainer from '../../containers/CommitItem'
-import Spinner from '../Spinner'
+import Spinner from '../chrome/Spinner'
 
 export default class Commits extends Base {
   constructor (props) {
