+++ lib/components/dataset/Dataset.js
@@ -12,7 +12,7 @@ import Render from './Render'
 import Export from './Export'
 import DatasetRouter from './DatasetRouter'
 
-import Spinner from '../Spinner'
+import Spinner from '../chrome/Spinner'
 
 // DATASET_DEBUG = 1 will print the if statements and functions run in componentWillMount and componentWillRecieveProps
 // DATASET_DEBUG = 2 will print the above and also the props, or the props and nextProps, in the componentWillMount and componentWillRecieveProps
