+++ lib/components/dataset/DatasetHeader.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import { withRouter } from 'react-router-dom'
 
 import DatasetRefProps from '../../propTypes/datasetRefProps'
-import Button from '../Button'
+import Button from '../chrome/Button'
 import DatasetName from '../DatasetName'
 import Header from '../Header'
 import Hash from '../Hash'
