+++ lib/components/dataset/Export.js
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import PropTypes from 'prop-types'
-import Button from '../Button'
+import Button from '../chrome/Button'
 import Base from '../Base'
 
 export default class Export extends Base {
