+++ lib/components/dataset/Overview.js
@@ -6,7 +6,7 @@ import Base from '../Base'
 import TagList from '../TagList'
 import StatsBar from '../StatsBar'
 import CommitItem from '../item/CommitItem'
-import Button from '../Button'
+import Button from '../chrome/Button'
 
 import { defaultPalette } from '../../propTypes/palette'
 
@@ -177,7 +177,7 @@ export default class Overview extends Base {
         {/* structure && updated && this.renderTitle(css, 'Structure', 'structure') */}
         {structure && updated && <StatsBar stats={this.stats(structure)} updated={updated} />}
         {/* commit && this.renderTitle(css, 'Commit', 'commit') */}
-        {commit && <CommitItem data={datasetRef} profile={profile} /> }
+        {commit && <CommitItem data={datasetRef} profile={profile} index={0} /> }
         {/* this.renderFields(md, css) */}
         <div className={css('buttonWrap')}><Button color='a' large full link={`${newUrl}/body`} text='TAKE ME TO THE DATA!' /></div>
       </div>
