+++ lib/components/dataset/Viz.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import Base from '../Base'
-import Spinner from '../Spinner'
+import Spinner from '../chrome/Spinner'
 import CodeViewer from '../CodeViewer'
 
 export default class Viz extends Base {
