+++ lib/components/editor/EditHeader.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import Button from '../Button'
+import Button from '../chrome/Button'
 import DatasetName from '../DatasetName'
 import Header from '../Header'
