+++ lib/components/editor/EditMeta.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 
 import Base from '../Base'
-import Button from '../Button'
+import Button from '../chrome/Button'
 import MetadataForm from '../form/MetadataForm'
 
 export default class EditMeta extends Base {
