+++ lib/components/editor/EditTransform.js
@@ -5,7 +5,7 @@ import MonacoEditor from 'react-monaco-editor'
 import { defaultPalette } from '../../propTypes/palette'
 
 import Base from '../Base'
-import Button from '../Button'
+import Button from '../chrome/Button'
 import KeyValueInput from '../form/KeyValueInput'
 
 const starter = [
