+++ lib/components/editor/EditViz.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import MonacoEditor from 'react-monaco-editor'
 
 import Base from '../Base'
-import Button from '../Button'
+import Button from '../chrome/Button'
 
 const starter = [
   '<!DOCTYPE html>',
