+++ lib/components/editor/Editor.js
@@ -13,8 +13,8 @@ import NavLinks from '../NavLinks'
 import { addActiveToLink } from '../../utils/links'
 
 import Base from '../Base'
-import Dropdown from '../Dropdown'
-import DropdownMenu from '../DropdownMenu'
+import Dropdown from '../chrome/Dropdown'
+import DropdownMenu from '../chrome/DropdownMenu'
 
 const RunOpts = [
   'Dry Run',
