+++ lib/components/form/SaveMetadataForm.js
@@ -5,8 +5,8 @@ import ValidInput from './ValidInput'
 import ValidTextarea from './ValidTextarea'
 
 import Base from '../Base'
-import Button from '../Button'
-import Spinner from '../Spinner'
+import Button from '../chrome/Button'
+import Spinner from '../chrome/Spinner'
 import { defaultPalette } from '../../propTypes/palette'
 
 export default class SaveMetadataForm extends Base {
