+++ lib/components/item/DatasetItem.js
@@ -7,7 +7,7 @@ import DatasetName from '../DatasetName'
 import Hash from '../Hash'
 import { defaultPalette } from '../../propTypes/palette'
 import Base from '../Base'
-import Spinner from '../Spinner'
+import Spinner from '../chrome/Spinner'
 
 export default class DatasetItem extends Base {
   constructor (props) {
