+++ lib/components/profile/Profile.js
@@ -8,7 +8,7 @@ import ProfileHeader from './ProfileHeader'
 import TabPanel from '../TabPanel'
 import List from '../List'
 import DatasetItem from '../item/DatasetItem'
-import Spinner from '../Spinner'
+import Spinner from '../chrome/Spinner'
 import NoProfile from './NoProfile'
 import ProfileSidebar from './ProfileSidebar'
 import ProfileEditorContainer from '../../containers/ProfileEditor'
@@ -31,7 +31,6 @@ export default class Profile extends Base {
       'renderProfileDatasets',
       'handleGoBack',
       'handleAddDataset',
-      'handleShowAddProfileDatasetModal',
       'handleSetProfilePoster',
       'handleSetProfilePhoto',
       'handleShowEditProfileModal',
@@ -62,7 +61,7 @@ export default class Profile extends Base {
     if (this.state.noProfile) {
       return
     }
-    if (!nextProps.peer && !nextProps.sessionProfileId) {
+    if (!nextProps.peer && this.props.sessionProfileId !== nextProps.sessionProfileId) {
       this.props.loadSessionProfile()
     } else if (this.props.peername !== nextProps.peername) {
       this.props.loadProfileByName(nextProps.peername).then((action) => {
@@ -145,7 +144,6 @@ export default class Profile extends Base {
           // onSelectItem={this.onSelectDataset}
           emptyComponent={<p>{this.handleMessage()}</p>}
           palette={palette}
-          onAdd={this.handleShowAddProfileDatasetModal}
         />
       )
     }
