+++ lib/components/profile/ProfileEditor.js
@@ -2,9 +2,9 @@ import React from 'react'
 import PropTypes from 'prop-types'
 
 import Base from '../Base'
-import Spinner from '../Spinner'
+import Spinner from '../chrome/Spinner'
 import ProfileForm from '../form/ProfileForm'
-import Button from '../Button'
+import Button from '../chrome/Button'
 import { Palette, defaultPalette } from '../../propTypes/palette'
 import ReadOnly from '../ReadOnly'
