+++ lib/components/profile/ProfileSidebar.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 
 import Base from '../Base'
-import Button from '../Button'
+import Button from '../chrome/Button'
 
 export default class ProfileSidebar extends Base {
   template (css) {
