+++ lib/components/stylesheet/Stylesheet.js
@@ -10,37 +10,20 @@ import DatasetName from '../DatasetName'
 import Datestamp from '../Datestamp'
 import StatsLine from '../StatsLine'
 import TabPanel from '../TabPanel'
-import Search from '../Search'
 import StatItem from '../item/StatItem'
-import Button from '../Button'
+import Button from '../chrome/Button'
 import List from '../List'
-import Dropdown from '../Dropdown'
-import DropdownMenu from '../DropdownMenu'
+import Dropdown from '../chrome/Dropdown'
+import DropdownMenu from '../chrome/DropdownMenu'
 import RadioInput from '../form/RadioInput'
 
 export default class Stylesheet extends Base {
   constructor (props) {
     super(props)
     this.state = {
-      searchString1: '',
-      searchString2: '',
-      searchString3: ''
     };
     [
-      'handleSearch1',
-      'handleSearch2',
-      'handleSearch3'
     ].forEach((m) => { this[m] = this[m].bind(this) })
-  }
-
-  handleSearch1 (searchString) {
-    this.setState({ searchString1: searchString })
-  }
-  handleSearch2 (searchString) {
-    this.setState({ searchString2: searchString })
-  }
-  handleSearch3 (searchString) {
-    this.setState({ searchString3: searchString })
   }
 
   template (css) {
@@ -106,9 +89,6 @@ export default class Stylesheet extends Base {
               <ValidInput type='text' name='input' label='valid input' value='input' showValidation={false} validation='blah' onChange={() => undefined} />
               <ValidInput type='text' name='input' label='invalid input' value='input' showValidation error='blah' onChange={() => undefined} />
               <Button color='a' text='Primary Button' name='primaryButton' />
-              <div><Search onChange={this.handleSearch1} searchString={this.state.searchString1} /></div>
-              <div><Search onChange={this.handleSearch2} searchString={this.state.searchString2} large /></div>
-              <div><Search onChange={this.handleSearch3} searchString={this.state.searchString3} border /></div>
             </div>
             <div className='col-md-6' >
               <DatasetName name='ramfox/comics' />
